Add k9s
K9s is a curses tool for interacting with kubernetes. It's very
popular amount the developers within the company and in the community.
It has been recently added to Arch (community/k9s), and as we trust the
Arch TUs we can now add this package. I've pinned it to the version that
is currently in [Arch Community](https://www.archlinux.org/packages/community/x86_64/k9s/)