Add SOPS to toolbox
With DO-927: Spin up a bastion and shell-host in CIDONE and DO-898: Move
secrets backup to terraformDONE finished, it’s now possible to run aws
`s3 sync s3://bitgo-ci-secrets/ secrets/` from the CI shell-host to
obtain the SOPS-encrypted secrets from S3.

Now, we need the remaining tooling in toolbox so that we can decrypt and
apply these secrets to the cluster. This requires, at a minimum, SOPS
itself.

Jira Ticket: DO-934